refactor(client): restructure video feature module for clarity
This refactoring simplifies the video feature by removing unnecessary nesting in the component
and page directories.

- Deleted `apps/client/src/features/video/components/ui`
- Deleted `apps/client/src/features/video/components/videos-dashboard`
- Moved and consolidated video-related components directly into
`apps/client/src/features/video/components`
- Renamed `videos-dashboard.page.tsx` to `videos-page.tsx` for consistency.

This change improves maintainability and makes the feature's structure more intuitive.
