--- apps/client/src/features/video/components/add-video-dialog/index.ts
@@ -1,1 +1,0 @@
-export * from "./add-video-dialog";
