+++ apps/client/src/features/video/components/index.ts
@@ -1,3 +1,2 @@
 export * from "./add-video-dialog";
-export * from "./ui";
-export * from "./videos-dashboard";
+export * from "./video-player";
