--- apps/client/src/features/video/components/ui/index.ts
@@ -1,5 +1,0 @@
-export * from "./video-note";
-export * from "./video-note-card";
-export * from "./video-notes-list";
-export * from "./video-page-header";
-export * from "./video-player";
