--- apps/client/src/features/video/components/ui/video-notes-list.tsx
@@ -1,31 +1,0 @@
-import type { Note } from "@tubenote/db";
-
-import { NotesListContainer } from "@/components/dashboards";
-
-import { VideoNoteCard } from "./";
-
-interface VideoNotesListProps {
-  notes: Note[];
-  setOpenMarkdownViewer: () => void;
-  setNote: (note: Note) => void;
-}
-
-export function VideoNotesList({
-  notes,
-  setOpenMarkdownViewer,
-  setNote,
-}: VideoNotesListProps) {
-  return (
-    <NotesListContainer notes={notes}>
-      {(note, onDeleteClick) => (
-        <VideoNoteCard
-          key={note.id}
-          note={note}
-          onDeleteClick={() => onDeleteClick(note.id)}
-          setOpenMarkdownViewer={setOpenMarkdownViewer}
-          setNote={setNote}
-        />
-      )}
-    </NotesListContainer>
-  );
-}
