+++ apps/client/src/features/video/pages/index.ts
@@ -1,2 +1,2 @@
 export * from "./video-page";
-export * from "./videos-dashboard.page";
+export * from "./videos-page";
