+++ apps/client/src/features/video/pages/videos-page/components/video-card/index.ts
