+++ apps/client/src/features/video/pages/videos-page/components/video-card/video-card-skeleton.tsx
