+++ apps/client/src/features/video/pages/videos-page/components/videos-list/videos-list-skeleton.tsx
