+++ apps/client/src/features/video/pages/videos-page/index.ts
@@ -1,0 +1,1 @@
+export * from "./videos-dashboard.page";
